# [SPARK-43232][SQL] Improve ObjectHashAggregateExec performance for high cardinality