<!--
Thanks for sending a pull request!  Here are some tips for you:
  1. If this is your first time, please read our contributor guidelines: https://spark.apache.org/contributing.html
  2. Ensure you have added or run the appropriate tests for your PR: https://spark.apache.org/developer-tools.html
  3. If the PR is unfinished, add '[WIP]' in your PR title, e.g., '[WIP][SPARK-XXXX] Your PR title ...'.
  4. Be sure to keep the PR description updated to reflect all changes.
  5. Please write your PR title to summarize what this PR proposes.
  6. If possible, provide a concise example to reproduce the issue for a faster review.
  7. If you want to add a new configuration, please read the guideline first for naming configurations in
     'core/src/main/scala/org/apache/spark/internal/config/ConfigEntry.scala'.
  8. If you want to add or modify an error type or message, please read the guideline first in
     'core/src/main/resources/error/README.md'.
-->

### What changes were proposed in this pull request?
<!--
Please clarify what changes you are proposing. The purpose of this section is to outline the changes and how this PR fixes the issue. 
If possible, please consider writing useful notes for better and faster reviews in your PR. See the examples below.
  1. If you refactor some codes with changing classes, showing the class hierarchy will help reviewers.
  2. If you fix some SQL features, you can provide some references of other DBMSes.
  3. If there is design documentation, please add the link.
  4. If there is a discussion in the mailing list, please add the link.
-->
The `ObjectHashAggregateExec` has three preformance issues:
- heavy overhead of scala sugar in `createNewAggregationBuffer`
- unnecessary grouping key comparation after fallback to sort based aggregator
- the aggregation buffer in sort based aggregator is not reused for all rest rows

Then the performance is poor with high cardinality grouping keys.

### Why are the changes needed?
<!--
Please clarify why the changes are needed. For instance,
  1. If you propose a new API, clarify the use case for a new API.
  2. If you fix a bug, you can clarify why it is a bug.
-->
Improve performance

### Does this PR introduce _any_ user-facing change?
<!--
Note that it means *any* user-facing change including all aspects such as the documentation fix.
If yes, please clarify the previous behavior and the change this PR proposes - provide the console output, description and/or an example to show the behavior difference if possible.
If possible, please also clarify if this is a user-facing change compared to the released Spark versions or within the unreleased branches such as master.
If no, write 'No'.
-->
no

### How was this patch tested?
<!--
If tests were added, say they were added here. Please make sure to add some test cases that check the changes thoroughly including negative and positive cases if possible.
If it was tested in a way different from regular unit tests, please clarify how you tested step by step, ideally copy and paste-able, so that other reviewers can test and check, and descendants can verify in the future.
If tests were not added, please describe why they were not added and/or why it was difficult to add.
If benchmark tests were added, please run the benchmarks in GitHub Actions for the consistent environment, and the instructions could accord to: https://spark.apache.org/developer-tools.html#github-workflow-benchmarks.
-->
The test should be covered by `org.apache.spark.sql.execution.aggregate.SortBasedAggregationStoreSuite`

Add benchmark for high cardinality case. Note, in this case the performance with no fallback is slower than others.

```sql
-- 10 * 1000 * 1000 rows
df.groupBy("key1", "key2")
  .agg(sum($"value"), count($"value"), avg($"value"), max($"value"), min($"value"),
    collect_set($"value"))
  .noop()
```

before
```
OpenJDK 64-Bit Server VM 1.8.0_322-b06 on Mac OS X 13.2
Apple M1 Pro
object agg v.s. sort agg with high cardinality:  Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
----------------------------------------------------------------------------------------------------------------------------------------------
sort agg                                                 4997           5099         118          2.0         499.7       1.0X
object agg w/o fallback                                 12634          12944         311          0.8        1263.4       0.4X
object agg w/ fallback                                   8792           8871          84          1.1         879.2       0.6X
```
after
```
OpenJDK 64-Bit Server VM 1.8.0_322-b06 on Mac OS X 13.2
Apple M1 Pro
object agg v.s. sort agg with high cardinality:  Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------------------------------------
sort agg                                                 4831           5060         199          2.1         483.1       1.0X
object agg w/o fallback                                  8971           9257         259          1.1         897.1       0.5X
object agg w/ fallback                                   5612           5706         102          1.8         561.2       0.9X
```